Refactor NoteCard component and add NotePage component
