+++ src/components/features/note/NoteCard/component.tsx
@@ -1,4 +1,3 @@
-import { Provider } from "jotai";
 import Link from "next/link";
 
 import { UserIcon } from "@/components/features/user/UserIcon";
@@ -10,7 +9,6 @@ import { DeleteButton } from "./DeleteButton";
 import { LikeButton } from "./LikeButton";
 import { PublishedAt } from "./PublishedAt";
 import { ReplyButton } from "./ReplyButton";
-import { ReplyForm } from "./ReplyForm";
 
 export type Props = {
   note: noteService.NoteCard;
@@ -18,40 +16,35 @@ export type Props = {
 
 export function NoteCard({ note }: Props) {
   return (
-    <Provider>
-      <Card
-        as="article"
-        data-testid="note-card"
-        className="text-prmary relative mb-1 space-y-2 p-4"
-      >
-        <div className="flex pl-[48px]">
-          <Link className="absolute left-3" href={note.user.url}>
-            <UserIcon user={note.user} className="rounded-full" size={36} />
-          </Link>
-          <div className="w-full space-y-2">
-            <div className="flex">
-              <Link href={note.user.url} className="truncate hover:underline">
-                {note.user.name && (
-                  <span className="mr-1 font-bold">{note.user.name}</span>
-                )}
-                <span className="text-gray">{note.user.displayUsername}</span>
-              </Link>
-              <PublishedAt href={note.url} publishedAt={note.publishedAt} />
-            </div>
-            <div dangerouslySetInnerHTML={{ __html: note.content }}></div>
-            {/* TODO: 複数枚画像に対応する */}
-            {note.attachmentUrl && (
-              <AttachmentImages url={note.attachmentUrl} />
-            )}
+    <Card
+      as="article"
+      data-testid="note-card"
+      className="text-prmary relative mb-1 space-y-2 p-4"
+    >
+      <div className="flex pl-[48px]">
+        <Link className="absolute left-3" href={note.user.url}>
+          <UserIcon user={note.user} className="rounded-full" size={36} />
+        </Link>
+        <div className="w-full space-y-2">
+          <div className="flex">
+            <Link href={note.user.url} className="truncate hover:underline">
+              {note.user.name && (
+                <span className="mr-1 font-bold">{note.user.name}</span>
+              )}
+              <span className="text-gray">{note.user.displayUsername}</span>
+            </Link>
+            <PublishedAt href={note.url} publishedAt={note.publishedAt} />
           </div>
+          <div dangerouslySetInnerHTML={{ __html: note.content }}></div>
+          {/* TODO: 複数枚画像に対応する */}
+          {note.attachmentUrl && <AttachmentImages url={note.attachmentUrl} />}
         </div>
-        <footer className="mt-3 flex gap-10 pl-[48px]">
-          <ReplyButton />
-          <LikeButton noteId={note.id} isLiked={note.isLiked} />
-          {note.isMine && <DeleteButton noteId={note.id} />}
-        </footer>
-      </Card>
-      <ReplyForm />
-    </Provider>
+      </div>
+      <footer className="mt-3 flex gap-10 pl-[48px]">
+        <ReplyButton url={note.url} />
+        <LikeButton noteId={note.id} isLiked={note.isLiked} />
+        {note.isMine && <DeleteButton noteId={note.id} />}
+      </footer>
+    </Card>
   );
 }
