+++ src/components/pages/NotePage/index.ts
